requirements.txt
flask
flask-socketio
opencv-python-headless
numpy
python-dotenv
Pillow
gtts
google-generativeai
gunicorn
eventlet
requests
werkzeug
